Updated the config for object detection
Code for training within the newest version of Tensorflow is altered and no longer recognizes:
schedule {
            step: 0
            learning_rate: .0003
          }
Therefore it is removed.
Also, the newest create_pet_tf_record.py script creates batches of files in th form of pet_faces_val.record-00000-of-00010 and so on. The config file is changed accordingly on line 121 and line 135 to be the corresponding line.